requirements.txt
Flask==3.0.3

uuid==1.30
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2